feat: add company service, move models
