+++ libs/shared/models/src/index.ts
@@ -1,3 +1,5 @@
 export * from './lib/auth/index';
 export * from './lib/aws/index';
+export * from './lib/company/index';
+export * from './lib/backend-responses/index';
 export * from './lib/environment';
