+++ libs/shared/models/src/lib/backend-responses/company-query.ts
@@ -1,0 +1,13 @@
+import { Company } from '../company';
+
+export type ListCompaniesQuery = {
+  listCompanies: ListCompaniesQueryResult;
+};
+
+export type ListCompaniesQueryResult = {
+  items: Company[];
+};
+
+export type CreateCompanyQuery = {
+  createCompany: Company;
+};
