+++ libs/shared/models/src/lib/backend-responses/error.ts
@@ -1,0 +1,13 @@
+export interface Error {
+  data: unknown;
+  errorInfo: string;
+  errorType: string;
+  locations: unknown;
+  message: string;
+  path: unknown;
+}
+
+export interface RequestError {
+  data: unknown;
+  errors: Error[];
+}
