+++ libs/shared/models/src/lib/backend-responses/index.ts
@@ -1,0 +1,2 @@
+export * from './error';
+export * from './company-query';
