+++ libs/shared/models/src/lib/company/company.ts
@@ -1,0 +1,8 @@
+export type Company = {
+  city: string;
+  companyName: string;
+  dotNumber: string;
+  state: string;
+  streetAddress: string;
+  zipCode: number;
+};
