+++ libs/shared/models/src/lib/company/index.ts
@@ -1,0 +1,1 @@
+export * from './company';
