--- libs/trucking-kit/company/data-access/src/lib/company-data-access.component.html
@@ -1,1 +1,0 @@
-<p>company-data-access works!</p>
