--- libs/trucking-kit/company/data-access/src/lib/company-data-access.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'tk-company-data-access',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './company-data-access.component.html',
-  styleUrl: './company-data-access.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class CompanyDataAccessComponent {}
